+++ pioreactor/actions/led_intensity.py
@@ -111,7 +111,7 @@ def led_intensity(
     pubsub_client: Optional[Client] = None,
 ) -> bool:
     """
-    Change the intensity of the LED channels A,B,C, or D to a intensity between 0 and 100.
+    Change the intensity of the LED channels A,B,C, or D to an value between 0 and 100.
 
     Parameters
     ------------