+++ CHANGELOG.md
@@ -3,6 +3,8 @@
  - Revisit pump and od calibrations. Some changes to the CLI.
  - Some improvements to stirring calibration.
  - Fix stirring calibration not responding in the UI
+ - Fix uninstall plugin bug
+ - Fix booleans not showing correctly in the PioreactorUI
 
 ### 22.10.4
  - removed the dbm storage `pio_jobs_running`. Instead, each job will write metadata about its currently running state to the (tmp) file `job_metadata_<job_name>`. This fixes issue #350.